rankingの順位の表示部分をコンポーネント化
